+++ src/test/java/wooteco/subway/AcceptanceTest.java
@@ -1,25 +1,28 @@
 package wooteco.subway;
 
-import io.restassured.RestAssured;
-import io.restassured.specification.RequestSpecification;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
+
+import io.restassured.RestAssured;
+import io.restassured.specification.RequestSpecification;
 import wooteco.subway.service.line.dto.LineDetailResponse;
 import wooteco.subway.service.line.dto.LineResponse;
 import wooteco.subway.service.line.dto.WholeSubwayResponse;
+import wooteco.subway.service.member.dto.LoginRequest;
 import wooteco.subway.service.member.dto.MemberResponse;
+import wooteco.subway.service.member.dto.TokenResponse;
 import wooteco.subway.service.path.dto.PathResponse;
 import wooteco.subway.service.station.dto.StationResponse;
-
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql("/truncate.sql")
@@ -44,13 +47,13 @@ public class AcceptanceTest {
     @LocalServerPort
     public int port;
 
+    public static RequestSpecification given() {
+        return RestAssured.given().log().all();
+    }
+
     @BeforeEach
     public void setUp() {
         RestAssured.port = port;
-    }
-
-    public static RequestSpecification given() {
-        return RestAssured.given().log().all();
     }
 
     public StationResponse createStation(String name) {
@@ -259,16 +262,18 @@ public class AcceptanceTest {
                         contentType(MediaType.APPLICATION_JSON_VALUE).
                         accept(MediaType.APPLICATION_JSON_VALUE).
                         when().
-                        post("/members").
+                        post("/members/signup").
                         then().
                         log().all().
                         statusCode(HttpStatus.CREATED.value()).
                         extract().header("Location");
     }
 
-    public MemberResponse getMember(String email) {
+    public MemberResponse getMember(String email, TokenResponse tokenResponse) {
         return
                 given().
+                        auth().
+                        oauth2(tokenResponse.getAccessToken()).
                         accept(MediaType.APPLICATION_JSON_VALUE).
                         when().
                         get("/members?email=" + email).
@@ -300,6 +305,33 @@ public class AcceptanceTest {
                 then().
                 log().all().
                 statusCode(HttpStatus.NO_CONTENT.value());
+    }
+
+    public MemberResponse myInfoWithBearerAuth(TokenResponse tokenResponse) {
+        return
+                given()
+                        .auth()
+                        .oauth2(tokenResponse.getAccessToken())
+                        .when()
+                        .get("/me/bearer")
+                        .then()
+                        .log().all()
+                        .statusCode(HttpStatus.OK.value())
+                        .extract().as(MemberResponse.class);
+    }
+
+    public TokenResponse login(String email, String password) {
+        return
+                given().
+                        body(new LoginRequest(TEST_USER_EMAIL, TEST_USER_PASSWORD)).
+                        contentType(MediaType.APPLICATION_JSON_VALUE).
+                        accept(MediaType.APPLICATION_JSON_VALUE).
+                        when().
+                        post("/oauth/token").
+                        then().
+                        log().all().
+                        statusCode(HttpStatus.OK.value()).
+                        extract().as(TokenResponse.class);
     }
 }
