+++ src/test/java/wooteco/subway/AuthAcceptanceTest.java
@@ -1,13 +1,10 @@
 package wooteco.subway;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
-import wooteco.subway.service.member.dto.LoginRequest;
 import wooteco.subway.service.member.dto.MemberResponse;
 import wooteco.subway.service.member.dto.TokenResponse;
 
@@ -22,32 +19,5 @@ public class AuthAcceptanceTest extends AcceptanceTest {
         assertThat(memberResponse.getId()).isNotNull();
         assertThat(memberResponse.getEmail()).isEqualTo(TEST_USER_EMAIL);
         assertThat(memberResponse.getName()).isEqualTo(TEST_USER_NAME);
-    }
-
-    public MemberResponse myInfoWithBearerAuth(TokenResponse tokenResponse) {
-        return
-            given()
-                .auth()
-                .oauth2(tokenResponse.getAccessToken())
-                .when()
-                .get("/me/bearer")
-                .then()
-                .log().all()
-                .statusCode(HttpStatus.OK.value())
-                .extract().as(MemberResponse.class);
-    }
-
-    public TokenResponse login(String email, String password) {
-        return
-            given().
-                body(new LoginRequest(TEST_USER_EMAIL, TEST_USER_PASSWORD)).
-                contentType(MediaType.APPLICATION_JSON_VALUE).
-                        accept(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                        post("/oauth/token").
-                then().
-                        log().all().
-                        statusCode(HttpStatus.OK.value()).
-                        extract().as(TokenResponse.class);
     }
 }
